refactor(daemon): extract plugin manifest + plugin-share helpers into plugin-share.ts
Strangler-fig slice 3: move readProjectPluginManifest, githubRepoNameFromPluginName,
PLUGIN_SHARE_ACTION_LABELS, USER_PLUGIN_SOURCE_KINDS, the plugin-context SKIP sets,
normalizePluginShareAction, renderPluginSharePrompt,
copyPluginFolderForProjectContext, copyPluginContextDir, shouldSkipPluginContextEntry
out of server.ts into a new sibling module, byte-identical.

All are pure or filesystem-only (no daemon module state). server.ts imports back
the six symbols it references and re-exports __forTestReadProjectPluginManifest to
preserve its public surface. normalizePluginShareAction's PLUGIN_SHARE_ACTION_PLUGIN_IDS
dep is a @open-design/contracts package import (no cycle).

No behavior change. Orphan/duplicate-binding checks clean.
